Refactor messages into individual classes
This will be useful when we start processing message commands.

For example, when we tell the server to start a Conference, we will receive back
an instantiated Conference object that contains the relevant IDs for that conference.
Then we will be able to do:

conf_obj.mute

which will generate a "mute" message to that specific conference.