Fix the build
Something in the switch to expect() syntax has caused #should not to be defined on all objects, leaving this to be called as a method on the actor. Once these call actors are dead, however, they can't respond to this. We need to explicitly look at #alive?

Also see https://github.com/adhearsion/adhearsion/commit/11346121832ce2b3d92fb41784bbc5baaa01418a

This is pending an upgrade to RSpec 3.x