Avoid races in processing Asterisk 13 bridges
On Asterisk 13, BridgeEnter and BridgeLeave events do not include the opposite-channel ID, leaving us to track this using the sequential events. When processed by the Call object, these must call back out to the translator to store global state to correlate. In previous releases of Punchblock, this has encountered a race condition because the translator is accessed via its ActorProxy, raising the possibility of the second event being processed before the first is done.

A previous attempt to fix this was https://github.com/adhearsion/punchblock/commit/14f6e760acbd8aac11f1410665bb5f6046db351a, released in 2.7.4, but this resulted in a suspected memory leak and crashing under heavy load in one example application. This first attempt was rolled back and the release yanked. This second attempt eliminates the issue by other means, that is moving the correlation up into the translator itself, passing the correlated data down to the call itself.